Response Message Body (Html, JSON)

+++ src/main/java/hello/servlet/basic/HelloData.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class HelloData {
 
     private String username;
-    private short age;
+    private int age;
 }
